requirements.txt
caldav==2.0.1
canvasapi==3.3.0
joblib==1.4.2
matplotlib==3.8.3
numpy==2.3.2
pandas==2.3.1
pathvalidate==3.3.1
python-dotenv==1.1.1
pytz==2024.1
Requests==2.32.4
scikit_learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
streamlit==1.39.0
urllib3==2.5.0
